+++ findmyself/src/main/java/com/findmyself/team/data/domain/SafetyCluster.java
@@ -9,12 +9,12 @@ import java.util.List;
 public class SafetyCluster {
 
     private int no;
-    private int min;
-    private int max;
+    private double min;
+    private double max;
     private double avg;
     private List<Long> codes;
 
-    public SafetyCluster(int no, int min, int max, double avg, List<Long> codes) {
+    public SafetyCluster(int no, double min, double max, double avg, List<Long> codes) {
         this.no = no;
         this.min = min;
         this.max = max;
