+++ findmyself/src/main/java/com/findmyself/team/data/domain/traffic/Traffic.java
@@ -18,9 +18,9 @@ public class Traffic {
     @GeneratedValue
     private Long h_code;
 
-    private int value;
+    private double value;
     private int no;
-    private int min;
-    private int max;
+    private double min;
+    private double max;
     private double avg;
 }
